refactor: rename integrationTest/ => testInteg/
Problem:
Integ tests live in a folder that has a different prefix than other test
directories, which reduces discoverability.

Solution:
Rename "integrationTest/" => "testInteg/". Now all tests live in
directories named `test*/`.
