+++ .nycrc.json
@@ -1,5 +1,5 @@
 {
     "all": true,
     "include": ["src/**", "dist/src/**"],
-    "exclude": ["test/**", "testE2E/**", "testLint/**", "integrationTest/**", "testFixtures/**"]
+    "exclude": ["test/**", "testE2E/**", "testLint/**", "testInteg/**", "testFixtures/**"]
 }
