+++ .vscode/launch.json
@@ -93,7 +93,7 @@
             "args": [
                 "${workspaceFolder}/dist/src/testFixtures/workspaceFolder",
                 "--extensionDevelopmentPath=${workspaceFolder}",
-                "--extensionTestsPath=${workspaceFolder}/dist/src/integrationTest/index.js"
+                "--extensionTestsPath=${workspaceFolder}/dist/src/testInteg/index.js"
             ],
             "env": {
                 "DEVELOPMENT_PATH": "${workspaceFolder}",
@@ -110,7 +110,7 @@
             "args": [
                 "${workspaceFolder}/dist/src/testFixtures/workspaceFolder",
                 "--extensionDevelopmentPath=${workspaceFolder}",
-                "--extensionTestsPath=${workspaceFolder}/dist/src/integrationTest/index.js"
+                "--extensionTestsPath=${workspaceFolder}/dist/src/testInteg/index.js"
             ],
             "env": {
                 "TEST_FILE": "${relativeFile}",
