+++ buildspec/linuxIntegrationTests.yml
@@ -68,7 +68,7 @@ phases:
     build:
         commands:
             - npm ci --unsafe-perm
-            - xvfb-run npm run integrationTest
+            - xvfb-run npm run testInteg
             - VCS_COMMIT_ID="${CODEBUILD_RESOLVED_SOURCE_VERSION}"
             - CI_BUILD_URL=$(echo $CODEBUILD_BUILD_URL | sed 's/#/%23/g')
             - CI_BUILD_ID="${CODEBUILD_BUILD_ID}"
