+++ package.json
@@ -3443,7 +3443,7 @@
         "testCompile": "npm run buildScripts && tsc -p ./ && npm run instrument",
         "test": "npm run testCompile && ts-node ./scripts/test/test.ts && npm run report",
         "testE2E": "npm run testCompile && ts-node ./scripts/test/testE2E.ts && npm run report",
-        "integrationTest": "npm run testCompile && ts-node ./scripts/test/integrationTest.ts && npm run report",
+        "testInteg": "npm run testCompile && ts-node ./scripts/test/testInteg.ts && npm run report",
         "lint": "ts-node ./scripts/lint/testLint.ts",
         "lintfix": "eslint -c .eslintrc.js --fix --ext .ts .",
         "package": "ts-node ./scripts/build/package.ts",
