+++ scripts/test/testInteg.ts
@@ -6,5 +6,5 @@
 import { resolve } from 'path'
 import { integrationSuite, runToolkitTests } from './launchTestUtilities'
 ;(async () => {
-    await runToolkitTests(integrationSuite, resolve('dist', 'src', 'integrationTest', 'index.js'))
+    await runToolkitTests(integrationSuite, resolve('dist', 'src', 'testInteg', 'index.js'))
 })()
