+++ src/test/cdk/explorer/appNode.test.ts
@@ -10,7 +10,7 @@ import * as vscode from 'vscode'
 import * as app from '../../../cdk/explorer/cdkProject'
 import * as appNode from '../../../cdk/explorer/nodes/appNode'
 import { ConstructNode } from '../../../cdk/explorer/nodes/constructNode'
-import { getTestWorkspaceFolder } from '../../../integrationTest/integrationTestsUtilities'
+import { getTestWorkspaceFolder } from '../../../testInteg/integrationTestsUtilities'
 import { getIcon } from '../../../shared/icons'
 import * as treeUtils from '../treeTestUtils'
 import { getLabel } from '../../../shared/treeview/utils'
