+++ src/test/codewhisperer/util/dependencyGraph/javascriptDependencyGraph.test.ts
@@ -7,7 +7,7 @@ import * as assert from 'assert'
 import * as vscode from 'vscode'
 import * as sinon from 'sinon'
 import * as fs from 'fs'
-import { getTestWorkspaceFolder } from '../../../../integrationTest/integrationTestsUtilities'
+import { getTestWorkspaceFolder } from '../../../../testInteg/integrationTestsUtilities'
 import { join } from 'path'
 import * as CodeWhispererConstants from '../../../../codewhisperer/models/constants'
 import { JavascriptDependencyGraph } from '../../../../codewhisperer/util/dependencyGraph/javascriptDependencyGraph'
