+++ src/testInteg/cloudformation/templateRegistry.test.ts
